+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+
+- Changed `layout` for application logging to log `X-Request-ID`
+
 ## [0.3.0] - 16-03-2020
 
 ### Added